tạo foreign key , tạo lớp đại diện interface, hoàn thiện chức năng CRUD nhân viên, cập nhật lại cơ sở dữ sdl

+++ database/migrations/2020_08_03_081142_create_hoadons_table.php
@@ -20,10 +20,10 @@ class CreateHoadonsTable extends Migration
             $table->time('giothanhtoan');
             $table->integer('ca');
             $table->timestamps();
-            $table->integer('id_giamgias');
-            $table->integer('id_vats');
-            $table->integer('id_bans');
-            $table->integer('id_users');
+            $table->unsignedBigInteger('id_giamgias');
+            $table->unsignedBigInteger('id_vats');
+            $table->unsignedBigInteger('id_bans');
+            $table->unsignedBigInteger('id_users');
             $table->softDeletes();
         });
     }
