+++ database/migrations/2020_08_03_090720_create_ctkhuyenmais_table.php
@@ -18,8 +18,8 @@ class CreateCtkhuyenmaisTable extends Migration
             $table->integer('phantramKM');
             $table->timestamps();
             $table->softDeletes();
-            $table->integer('id_khuyenmais');
-            $table->integer('id_mons');
+            $table->unsignedBigInteger('id_khuyenmais');
+            $table->unsignedBigInteger('id_mons');
         });
     }
